requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
dj-rest-auth==4.0.1
drf-spectacular==0.26.3
psycopg[binary]
sqlparse==0.4.4
django-crispy-forms
crispy-bootstrap5
django-allauth==0.54.0
environs[django]
pillow
